+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+## [0.5.4] - 2025-12-08
+
 ### Added
 
 - Type shortcuts for `typelize` method. ([@skryukov])
@@ -277,7 +279,8 @@ and this project adheres to [Semantic Versioning].
 [@prog-supdex]: https://github.com/prog-supdex
 [@ventsislaf]: https://github.com/ventsislaf
 
-[Unreleased]: https://github.com/skryukov/typelizer/compare/v0.5.3...HEAD
+[Unreleased]: https://github.com/skryukov/typelizer/compare/v0.5.4...HEAD
+[0.5.4]: https://github.com/skryukov/typelizer/compare/v0.5.3...v0.5.4
 [0.5.3]: https://github.com/skryukov/typelizer/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/skryukov/typelizer/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/skryukov/typelizer/compare/v0.5.0...v0.5.1